+++ common/src/client/java/io/github/jamalam360/jamlib/client/gui/ScrollingStringWidget.java
@@ -1,5 +1,6 @@
 package io.github.jamalam360.jamlib.client.gui;
 
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.StringWidget;
@@ -13,5 +14,9 @@ public class ScrollingStringWidget extends StringWidget {
 	@Override
 	public void renderWidget(GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) {
 		this.renderScrollingString(guiGraphics, this.getFont(), 2, this.getColor());
+
+		if (this.isMouseOver(mouseX, mouseY) && this.getTooltip() != null) {
+			guiGraphics.renderTooltip(Minecraft.getInstance().font, this.getTooltip().toCharSequence(Minecraft.getInstance()), mouseX, mouseY);
+		}
 	}
 }