+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Switch to fabric-api mod ID in dependencies block (closes #10).
 - Enable split source sets (closes #14).
 - Identify config managers by `(MOD_ID, CONFIG_NAME)` rather than by just `(CONFIG_NAME)` (closes #15).
+- Allow `List<?>` config fields (closes #11).